+++ spec/active_interaction/base_spec.rb
@@ -148,6 +148,28 @@ describe ActiveInteraction::Base do
         expect(outcome).to be_a described_class
       end
 
+      context 'setting the result' do
+        let(:described_class) do
+          Class.new(ActiveInteraction::Base) do
+            boolean :attribute
+
+            validate do
+              @_interaction_result = SecureRandom.hex
+              errors.add(:attribute, SecureRandom.hex)
+            end
+
+            def self.name
+              SecureRandom.hex
+            end
+          end
+        end
+
+        it 'sets the result to nil' do
+          expect(outcome).to be_invalid
+          expect(outcome.result).to be_nil
+        end
+      end
+
       context 'failing validations' do
         it 'returns an invalid outcome' do
           expect(outcome).to be_invalid
@@ -160,6 +182,27 @@ describe ActiveInteraction::Base do
 
       context 'passing validations' do
         before { options.merge!(thing: thing) }
+
+        context 'failing runtime validations' do
+          before do
+            @execute = described_class.instance_method(:execute)
+            described_class.send(:define_method, :execute) do
+              errors.add(:thing, SecureRandom.hex)
+            end
+          end
+
+          after do
+            described_class.send(:define_method, :execute, @execute)
+          end
+
+          it 'returns an invalid outcome' do
+            expect(outcome).to be_invalid
+          end
+
+          it 'sets the result to nil' do
+            expect(outcome.result).to be_nil
+          end
+        end
 
         it 'returns a valid outcome' do
           expect(outcome).to be_valid